site is running off of server/server.js and can be run with run develop in main folder
