+++ client/.env
@@ -1,0 +1,1 @@
+REACT_APP_BACKEND_URL = "http://localhost:3001"
